pyproject.toml
[project]
name = "weather-display"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "flask>=3.1.2",
    "pillow>=12.0.0",
    "requests>=2.32.5",
    "rich>=14.2.0",
]
